requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
openai==1.6.1
nltk==3.8.1
textstat==0.7.3
pdfplumber==0.10.3
Werkzeug==3.0.1
gunicorn==21.2.0
